requirements.txt
Django>=4.2,<5.2
openai>=1.0
google-genai>=1.0
groq>=0.4
